Estimate Inverse Probability of Censoring Weights

Estimate Inverse Probability of Censoring Weights

est_ipcw(V, Delta, fit_type = c("sl", "glm"), sl_learners = NULL)

Arguments

V
A numeric vector, matrix, data.frame or similar object giving the observed values of the covariates known to potentially inform the censoring mechanism.

Delta
A numeric vector giving observed values of the indicator function corresponding to the censoring mechanism.

fit_type
A character indicating whether to perform the fit using GLMs or a Super Learner. If use of Super Learner is desired, then the argument sl_learners must be provided.

sl_learners
An Lrnr_sl object, a Super Learner instantiated externally using sl3.

Value

A list containing a numeric vector corresponding to the inverse probability of censoring weights required for computing an IPCW-TMLE and numeric vector of the estimated missingness mechanism. Formally, the former is nothing more than term class as computed using standard logistic regression.

Details

Compute inverse probability of censoring weights for the two-phase sampling mechanism. These inverse weights are based on the probability of appearing in the second-phase sample based on variables measured on all participants.