Iterative Computation of IPCW-TMLE

An adaptation of the general IPCW-TMLE formulation of Rose & van der Laan as well as its associated algorithm. This may be used to iteratively construct an efficient IPCW-TMLE, which is computed by applying IPC weights to the efficient influence function of the parameter and subsequently updating the IPC weights. The efficient IPCW-TMLE estimate is provided by looping over this function until convergence criteria defined over the efficient influence function are satisfied (e.g., a mean of very nearly zero). For a detailed description of the IPCW-TMLE, please consult Rose and van der Laan (2011) <doi:10.2202/1557-4679.1217>. INTERNAL USE ONLY.

ipcw_tmle_proc(data_in, C, V, ipc_mech, ipc_weights, ipc_weights_norm,
  Qn_estim, Hn_estim, fluc_method = c("standard", "weighted"),
  fit_type = c("glm", "sl", "fit_spec"), eif_tol = 1e-09,
  sl_lrnrs = NULL, eif_reg_type = c("hal", "glm"))

Arguments

data_in
A data.table containing variables and observations that represent the fully observed data. That is, this is a table of the data as it is seen after a potential censoring process is applied.

C
A numeric binary vector giving the censoring status of a given observation.

V
A data.table giving the values across all observations of all variables that play a role in the censoring mechanism.

ipc_mech
A numeric vector containing values that describe the censoring mechanism for all of the observations. Note that such values are estimated by regressing the censoring covariates V on the observed censoring C and thus correspond to predicted probabilities of being censored for each observation.

ipc_weights
A numeric vector of inverse probability of censoring weights. These are equivalent to C / ipc_mech in any initial run of this function. Updated values of this vector are provided as part of the output of this function, which may be used in subsequent calls that allow convergence to a more efficient estimate.

ipc_weights_norm
A numeric vector of the same weights described in the previous argument, except normalized in this case.

Qn_estim
A data.table corresponding to the outcome regression. This is produced by invoking the internal function est_Q.

Hn_estim
A data.table corresponding to values produced in the computation of the auxiliary ("clever") covariate. This is produced easily by invoking the internal function est_Hn.

fluc_method
A character giving the type of regression to be used in traversing the fluctuation submodel. The choices are "weighted" and "standard" -- please consult the literature for details on the differences.

fit_type
A character providing the type of regression to be fit in estimating the relationship between the variables composing the censoring mechanism and the efficient influence function. Choose "glm" for generalized linear models or "sl" for use of the Super Learner algorithm. If selecting the latter, the final argument sl_lrnrs must be provided.

eif_tol
A numeric providing the largest value to be tolerated as the mean of the efficient influence function.

sl_lrnrs
A Lrnr_sl object composed of a collection of learners provided by the sl3 package. This is constructed externally from this function and provided as input.

eif_reg_type
Whether a flexible nonparametric function ought to be used in the dimension-reduced nuisance regression of the targeting step for the censored data case. By default, the method used is a nonparametric regression based on the Highly Adaptive Lasso (from package hal9001). Set this to "glm" to instead use a simple linear regression model. In this step, the efficient influence function (EIF) is regressed against covariates contributing to the censoring mechanism (i.e., EIF ~ V | C = 1).