Compute IPCW Part of the Efficient Influence Function for IPCW-TMLE

Computes an additional component of the efficient influence function needed for efficient estimation of IPCW-TMLEs. This takes the form:

ipcwtmle_eif(fluc_fit_out, Hn, Y, C, ipc_weights = rep(1, length(Y)),
  tol_eif = 1e-07)

Arguments

fluc_fit_out
...

Hn
...

Y
A numeric vector of observed outcomes.

ipc_weights
...

tol_eif
...