Summary for Counterfactual Mean of Stochastic Shift Intervention

Summary for Counterfactual Mean of Stochastic Shift Intervention

# S3 method for txshift
summary(object, ..., ci_level = 0.95, digits = 4)

Arguments

object
An object of class txshift, as produced by invoking the function txshift, for which a confidence interval is to be computed.

...
Other arguments. Not currently used.

ci_level
A numeric indicating the level of the confidence interval to be computed.

digits
A numeric scalar giving the number of digits to be displayed or to round results to.

Value

None. Called for the side effect of printing a summary of particular slots of objects of class txshift.

Details

Print a convenient summary for objects computed using txshift.

Examples

set.seed(429153)
n_obs <- 100
W <- replicate(2, rbinom(n_obs, 1, 0.5))
A <- rnorm(n_obs, mean = 2 * W, sd = 1)
Y <- rbinom(n_obs, 1, plogis(A + W + rnorm(n_obs, mean = 0, sd = 1)))
txout <- txshift(
  W = W, A = A, Y = Y, delta = 0.5,
  estimator = "tmle",
  g_exp_fit_args = list(
    fit_type = "hal", n_bins = 5,
    grid_type = "equal_mass",
    lambda_seq = exp(-1:-9)
  ),
  Q_fit_args = list(
    fit_type = "glm",
    glm_formula = "Y ~ ."
  )
)
summary(txout)

#>     lwr_ci  param_est     upr_ci  param_var   eif_mean  estimator     n_iter 
#>     0.6967     0.7894     0.8595     0.0017 1.5266e-12       tmle          0