TML Estimate of the Effect of a Continuous Treatment

TML Estimate of the Effect of a Continuous Treatment

tmle_shift(Y, A, W, Qn, gn, delta, tol = 1e-05, iter_max = 5, A_val)

Arguments

Y
A numeric vector of observed outcomes.

A
A numeric vector of observed treatments.

W
A matrix or data.frame of baseline covariates.

Qn
Function to compute the outcome regression: Q(A, W) = E(Y | A, W).

gn
Function to compute the propensity score: g(A, W) = density(A | W).

delta
A numeric for the shift to be placed on the treatment of interest (i.e., the effect of shifting treatment A by delta units).

tol
A numeric for the tolerance for measuring convergence of the parametric fluctuation.

iter_max
A numeric for the maximum number of iterations.

A_val
A vector of numeric values for the points in the range of the treatment A to approximate integrals by Riemmann sums. Note that these must be equally spaced along a grid.

Examples

n <- 100
W <- data.frame(W1 = runif(n), W2 = rbinom(n, 1, 0.7))
A <- rpois(n, lambda = exp(3 + 0.3 * log(W$W1) - 0.2 * exp(W$W1) * W$W2))
Y <- rbinom(n, 1, plogis(-1 + 0.05 * A - 0.02 * A * W$W2 +
                         0.2 * A * tan(W$W1^2) - 0.02 * W$W1 * W$W2 +
                         0.1 * A * W$W1 * W$W2))
fitA.0 <- glm(A ~ I(log(W1)) + I(exp(W1)):W2, family = poisson,
              data = data.frame(A, W))
fitY.0 <- glm(Y ~ A + A:W2 + A:I(tan(W1^2)) + W1:W2 + A:W1:W2,
              family = binomial, data = data.frame(A, W))
gn.0  <- function(A = A, W = W) {
  dpois(A, lambda = predict(fitA.0, newdata = W, type = "response"))
}
Qn.0 <- function(A = A, W = W) {
  predict(fitY.0, newdata = data.frame(A, W, row.names = NULL),
          type = "response")
}
tmle00 <- tmle_shift(Y = Y, A = A, W = W, Qn = Qn.0, gn = gn.0, delta = 2,
                     tol = 1e-4, iter_max = 5, A_val = seq(1, 60, 1))