Compute the Shift Parameter Estimate and the Efficient Influence Function

Compute the Shift Parameter Estimate and the Efficient Influence Function

eif(
  Y,
  Qn,
  Hn,
  estimator = c("tmle", "onestep"),
  fluc_mod_out = NULL,
  Delta = rep(1, length(Y)),
  ipc_weights = rep(1, length(Y)),
  ipc_weights_norm = rep(1, length(Y))
)

Arguments

Y
A numeric vector of the observed outcomes.

Qn
An object providing the value of the outcome evaluated after imposing a shift in the treatment. This object is passed in after being constructed by a call to the internal function est_Q.

Hn
An object providing values of the auxiliary ("clever") covariate, constructed from the treatment mechanism and required for targeted minimum loss-based estimation. This object object should be passed in after being constructed by a call to the internal function est_Hn.

estimator
The type of estimator to be fit, either "tmle" for targeted maximum likelihood estimation or "onestep" for a one-step augmented inverse probability weighted (AIPW) estimator.

fluc_mod_out
An object giving values of the logistic tilting model for targeted minimum loss estimation. This type of object should be the output of the internal routines to perform this step of the TML estimation procedure, as given by fit_fluctuation.

Delta
Indicator for missingness. Used for compatibility with the routine to compute IPCW-TML estimates.

ipc_weights
A numeric vector that gives inverse probability of censoring weights for each observation. These are generated by invoking the routines for estimating the censoring mechanism.

ipc_weights_norm
A numeric vector of the weights described in the previous argument. In this case, the weights are normalized.

Value

TODO

Details

Estimate the value of the causal parameter alongside statistical inference for the parameter estimate based on the efficient influence function of the target parameter, which takes the following form: