Marginal structural model for the causal effects of a grid of interventions

Marginal structural model for the causal effects of a grid of interventions

msm_vimshift(Y, A, W, C = rep(1, length(Y)), V = NULL,
  delta_grid = seq(-0.5, 0.5, 0.5), estimator = c("tmle", "onestep"),
  weighting = c("identity", "variance"), ci_level = 0.95,
  ci_type = c("marginal", "simultaneous"), ...)

Arguments

Y
A numeric vector of the observed outcomes.

A
A numeric vector corresponding to a treatment variable. The parameter of interest is defined as a location shift of this quantity.

W
A matrix, data.frame, or similar corresponding to a set of baseline covariates.

C
A numeric binary vector giving information on whether a given observation was subject to censoring. This is used to compute an IPCW-TMLE in cases where two-stage sampling is performed. The default assumes that no censoring was present (i.e., a two-stage design was NOT used). N.B., this is equivalent to the term van der Laan manuscript that introduced/formulated IPCW-TML estimators.

V
The covariates that are used in determining the sampling procedure that gives rise to censoring. The default is NULL and corresponds to scenarios in which there is no censoring (in which case all values in the preceding argument C must be uniquely 1. To specify this, pass in a NAMED list identifying variables amongst W, A, Y that are thought to have played a role in defining the sampling/censoring mechanism (C).

delta_grid
A numeric vector giving the individual values of the shift parameter used in computing each of the TML estimates.

estimator
The type of estimator to be fit, either "tmle" for targeted maximum likelihood estimation or "onestep" for a one-step augmented inverse probability weighted (AIPW) estimator.

weighting
Whether to weight each parameter estimate by the inverse of its variance (in order to improve stability of the resultant MSM fit) or to simply weight all parameter estimates equally. The default is the option "identity", weighting all estimates identically.

ci_level
A numeric indicating the desired coverage level of the confidence interval to be computed.

ci_type
Whether to construct a simultaneous confidence band covering all parameter estimates at once or marginal confidence intervals covering each parameter estimate separately. The default is to construct marginal confidence intervals for each parameter estimate rather than a simultaneous confidence band.

...
Additional arguments to be passed to txshift.

Examples

set.seed(429153)
n_obs <- 100
W <- as.numeric(replicate(1, rbinom(n_obs, 1, 0.5)))
A <- as.numeric(rnorm(n_obs, mean = 2 * W, sd = 1))
Y <- rbinom(n_obs, 1, plogis(A + W + rnorm(n_obs, mean = 0, sd = 1)))
msm <- msm_vimshift(
  W = W, A = A, Y = Y, estimator = "tmle",
  g_fit_args = list(
    fit_type = "hal",
    n_bins = 5,
    grid_type = "equal_mass",
    lambda_seq = exp(seq(-1, -9,
      length = 100
    ))
  ),
  Q_fit_args = list(
    fit_type = "glm",
    glm_formula = "Y ~ ."
  )
)