Compute Targeted Minimum Loss-Based Estimate of Counterfactual Mean Under Stochastically Shifted Treatment

This is the primary user-facing wrapper function to be used in invoking the procedure to obtain targeted maximum likelihood / targeted minimum loss-based estimates (TMLE) of the causal parameter discussed in Díaz and van der Laan (2013) and Díaz and van der Laan (2018).

tmle_txshift(data_internal, C = rep(1, nrow(data_internal)), V = NULL,
  delta, ipcw_estim, Qn_estim, Hn_estim, fluctuation = c("standard",
  "weighted"), eif_tol = 1/nrow(data_internal), max_iter = 10000,
  eif_reg_type = c("hal", "glm"), ipcw_fit_args,
  ipcw_efficiency = TRUE)

Arguments

data_internal
A data.table constructed internally by a call to txshift. This contains most of the data elements needed for computing the targeted maximum likelihood estimator.

C
A numeric binary vector giving information on whether a given observation was subject to censoring, used to compute an IPCW-TMLE in cases where two-stage sampling is performed. Default assumes no censoring.

V
The covariates that are used in determining the sampling procedure that gives rise to censoring. The default is NULL and corresponds to scenarios in which there is no censoring (in which case all values in the preceding argument C must be uniquely 1. To specify this, pass in a NAMED list identifying variables amongst W, A, Y that are thought to have played a role in defining the sampling/censoring mechanism (C).

delta
A numeric value indicating the shift in the treatment to be used in defining the target parameter. This is defined with respect to the scale of the treatment (A).

ipcw_estim
An object providing the value of the censoring mechanism evaluated across the full data. This object is passed in after being constructed by a call to the internal function est_ipcw.

Qn_estim
An object providing the value of the outcome evaluated after imposing a shift in the treatment. This object is passed in after being constructed by a call to the internal function est_Q.

Hn_estim
An object providing values of the auxiliary ("clever") covariate, constructed from the treatment mechanism and required for targeted minimum loss-based estimation. This object object should be passed in after being constructed by a call to the internal function est_Hn.

fluctuation
The method to be used in the submodel fluctuation step (targeting step) to compute the TML estimator. The choices are "standard" and "weighted" for where to place the auxiliary covariate in the logistic tilting regression.

eif_tol
A numeric giving the convergence criterion for the TML estimator. This is the maximum mean of the efficient influence function to be used in declaring convergence.

max_iter
A numeric integer giving the maximum number of steps to be taken in iterating to a solution of the efficient influence function.

eif_reg_type
Whether a flexible nonparametric function ought to be used in the dimension-reduced nuisance regression of the targeting step for the censored data case. By default, the method used is a nonparametric regression based on the Highly Adaptive Lasso (from package hal9001). Set this to "glm" to instead use a simple linear regression model. In this step, the efficient influence function (EIF) is regressed against covariates contributing to the censoring mechanism (i.e., EIF ~ V | C = 1).

ipcw_fit_args
A list of arguments, all but one of which are passed to est_ipcw. For details, please consult the documentation for est_ipcw. The first element of this (i.e., fit_type) is used to determine how this regression is fit: "glm" for generalized linear model, "sl" for a Super Learner, and "fit_spec" a user-specified input of the form produced by est_ipcw. NOTE THAT this first argument is not passed to est_ipcw.

ipcw_efficiency
Whether to invoke an augmentation of the IPCW-TMLE procedure that performs an iterative process to ensure efficiency of the resulting estimate. The default is TRUE; set to FALSE to use an IPC-weighted loss rather than the IPC-augment influence function.

Value

S3 object of class txshift containing the results of the procedure to compute a TML estimate of the treatment shift parameter.